# [WIP] Add fast cuda kernels for one mode