TODO: support seq mode

Add fast gemv kernel (based on https://github.com/wangsiping97/FastGEMV and added kernel fusion. ~10% faster than pytorch gemv) and a fused wkv_forward_one kernel (much faster, ~140 us -> 5us on 2080 1.5B model).

-------

**2080, 1.5B Model:**

Main branch (blue bars represent CUDA kernels):
<img width="723" alt="image" src="https://github.com/BlinkDL/ChatRWKV/assets/11607199/5cab0f6d-551a-44ea-8a47-5704108cb5a6">

This branch:

<img width="494" alt="image" src="https://github.com/BlinkDL/ChatRWKV/assets/11607199/ca6c5b47-cea3-409b-a640-bd9d79df42d3">

FFN time: 359.77us -> 231.29us
ATT time: 291.71us -> 148.45us

Whole one mode time: 0.2487s -> 0.1409s

**A100, 7B model:**

Whole one mode time: 0.2657s -> 0.184s